pyproject.toml
[tool.poetry]
name = "dash-hangout-example"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
dash = "3.0.4"
pandas = "^2.2.3"
spacy = "^3.8.5"
matplotlib = "^3.10.1"
sqlalchemy = "^2.0.40"
python-dotenv = "^1.1.0"
asyncpg = "^0.30.0"
dash-mantine-components = "1.2.0"
dash-iconify = "^0.1.2"
dash-flash = "^1.0.0"
nbformat = "^5.10.4"
greenlet = "^3.2.0"
kaleido = "1.0.0rc11"
dash-ag-grid = "^31.3.1"
flash-echarts = {path = "assets/flash_echarts-0.0.1.tar.gz"}
redis = "^5.2.1"
uvicorn = "0.34.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
